server/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autopep8==2.0.4
certifi==2023.11.17
click==8.1.7
distro==1.9.0
Django==4.2.9
docstring-parser==0.15
exceptiongroup==1.2.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
instructor==0.4.6
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.4
openai==1.7.1
pdf2image==1.17.0
pillow==10.2.0
pycodestyle==2.11.1
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
pypdfium2==4.26.0
rich==13.7.0
sniffio==1.3.0
sqlparse==0.4.4
tomli==2.0.1
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.9.0
yarl==1.9.4
